Switch to yaml safe_load and allow the use of YAML aliases
The bundled YAML intepreter bundled with Ruby changed the default
behaviour of the load() function from Ruby 3.1.

load() is no longer an alias for unsafe_load() and now defaults to
safe_load(), but this function prevents the use of YAML aliases.

For more info: https://github.com/ruby/psych/issues/533

As a reasonable compromise, this commit modifies the load behavior
to explicitly use the safe_load() function, but also enabled the
use of aliases.